+++ src/main/java/io/fabric8/maven/docker/SourceMojo.java
@@ -16,8 +16,11 @@ package io.fabric8.maven.docker;/*
  */
 
 import java.io.File;
+import java.util.ArrayList;
+import java.util.List;
 
 import io.fabric8.maven.docker.config.BuildImageConfiguration;
+import io.fabric8.maven.docker.config.SourceMode;
 import io.fabric8.maven.docker.util.MojoParameters;
 import io.fabric8.maven.docker.access.DockerAccessException;
 import io.fabric8.maven.docker.config.ImageConfiguration;
@@ -26,40 +29,85 @@ import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.*;
 import org.apache.maven.project.MavenProjectHelper;
 
+import static io.fabric8.maven.docker.config.RunImageConfiguration.NamingStrategy.alias;
+import static io.fabric8.maven.docker.config.WatchMode.build;
+import static org.apache.maven.plugin.assembly.utils.ProjectUtils.getClassifier;
+
 /**
  * Mojo for attaching one more source docker tar file to an artifact.
+ *
+ * If used in a lifecycle called 'docker-tar' the artifacts are attached
+ * without classifier. Otherwise the classifier is "docker" and, depending on the
+ * selected {@link SourceMode}, the alias name of the build configuration.
  *
  * @author roland
  * @since 25/10/15
  */
-@Mojo(name = "source", defaultPhase = LifecyclePhase.GENERATE_SOURCES)
-@Execute(phase = LifecyclePhase.GENERATE_SOURCES)
+@Mojo(name = "source", defaultPhase = LifecyclePhase.PACKAGE)
+@Execute(phase = LifecyclePhase.PACKAGE)
 public class SourceMojo extends  AbstractBuildSupportMojo {
 
     @Component
     private MavenProjectHelper projectHelper;
 
+    /**
+     * Mode how to attach the source artifact:
+     *
+     * <ul>
+     *     <li><strong>first</strong> : The first image with build configuration</li>
+     *     <li><strong>all</strong> : All images with build configuration. Each image must have an alias
+     *        configured which is use as part of the classifier.
+     *     </li>
+     * </ul>
+     */
+    @Parameter
+    private SourceMode sourceMode = SourceMode.first;
+
     @Override
     protected void executeInternal(ServiceHub hub) throws DockerAccessException, MojoExecutionException {
         MojoParameters params = createMojoParameters();
+        List<ImageConfiguration> imageConfigs = new ArrayList<>();
         for (ImageConfiguration imageConfig : getImages()) {
             BuildImageConfiguration buildConfig = imageConfig.getBuildConfiguration();
             if (buildConfig != null) {
                 if (buildConfig.skip()) {
                     log.info(imageConfig.getDescription() + ": Skipped creating source");
                 } else {
-                    File dockerTar =
-                            hub.getArchiveService().createDockerBuildArchive(imageConfig, params);
-                    String alias = imageConfig.getAlias();
-                    if (alias == null) {
-                        throw new IllegalArgumentException(
-                                "Image " + imageConfig.getDescription() + " must have an 'alias' configured to be " +
-                                "used as a classifier for attaching a docker build tar as source to the maven build");
-                    }
-                    projectHelper.attachArtifact(project, buildConfig.getCompression().getFileSuffix(),"docker-" + alias, dockerTar);
+                    imageConfigs.add(imageConfig);
                 }
             }
         }
+        if (sourceMode == SourceMode.first && imageConfigs.size() > 0) {
+            ImageConfiguration imageConfig = imageConfigs.get(0);
+            File dockerTar = hub.getArchiveService().createDockerBuildArchive(imageConfig, params);
+            projectHelper.attachArtifact(project, getArchiveType(imageConfig),
+                                         getClassifier(null), dockerTar);
+        } else {
+            for (ImageConfiguration imageConfig : imageConfigs) {
+                File dockerTar = hub.getArchiveService().createDockerBuildArchive(imageConfig, params);
+                String alias = imageConfig.getAlias();
+                if (alias == null) {
+                    throw new IllegalArgumentException(
+                        "Image " + imageConfig.getDescription() + " must have an 'alias' configured to be " +
+                        "used as a classifier for attaching a docker build tar as source to the maven build");
+                } else {
+                    projectHelper.attachArtifact(project, getArchiveType(imageConfig), getClassifier(alias), dockerTar);
+                }
+            }
+        }
+    }
+
+    private String getClassifier(String alias) {
+        String packaging = project.getPackaging();
+        if ("docker-tar".equalsIgnoreCase(packaging)) {
+            return alias;
+        } else {
+            return "docker" + (alias != null ? "-" + alias : "");
+        }
+    }
+
+    private String getArchiveType(ImageConfiguration imageConfig) {
+        return imageConfig.getBuildConfiguration().getCompression().getFileSuffix();
     }
 
     @Override